[#273] ref: rename KJsonTransformerBasedAssetLoader -> KJsonAssetLoader
