+++ implementations/implementations-std-core/src/test/java/io/github/darthakiranihil/konna/core/io/KJsonAssetLoaderPositiveTests.java
@@ -19,12 +19,15 @@ package io.github.darthakiranihil.konna.core.io;
 import io.github.darthakiranihil.konna.core.data.json.KJsonValue;
 import io.github.darthakiranihil.konna.core.io.protocol.KClasspathProtocol;
 import io.github.darthakiranihil.konna.test.KStandardTestClass;
+import org.jspecify.annotations.NullMarked;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
+import java.util.Objects;
 
-public class KJsonTransformerBasedAssetLoaderPositiveTests extends KStandardTestClass {
+@NullMarked
+public class KJsonAssetLoaderPositiveTests extends KStandardTestClass {
 
     private static final class Alias1Typedef implements KAssetTypedef {
 
@@ -54,7 +57,7 @@ public class KJsonTransformerBasedAssetLoaderPositiveTests extends KStandardTest
     @Test
     public void testLoadAsset() {
 
-        KAssetLoader assetLoader = new KJsonTransformerBasedAssetLoader(
+        KAssetLoader assetLoader = new KJsonAssetLoader(
             new KStandardResourceLoader(
                 List.of(new KClasspathProtocol(
                     ClassLoader.getSystemClassLoader()
@@ -85,7 +88,7 @@ public class KJsonTransformerBasedAssetLoaderPositiveTests extends KStandardTest
         );
 
         try {
-            assetLoader.addAssetTypedef(new KJsonTransformerBasedAssetLoaderPositiveTests.Alias1Typedef());
+            assetLoader.addAssetTypedef(new KJsonAssetLoaderPositiveTests.Alias1Typedef());
 
             KAsset asset = assetLoader.loadAsset("type_1.asset_1", "alias_1");
 
@@ -108,7 +111,7 @@ public class KJsonTransformerBasedAssetLoaderPositiveTests extends KStandardTest
             int[] intArray = asset.getIntArray("int_array_property");
             float[] floatArray = asset.getFloatArray("float_array_property");
             boolean[] booleanArray = asset.getBooleanArray("boolean_array_property");
-            String[] stringArray = asset.getStringArray("string_array_property");
+            String[] stringArray = Objects.requireNonNull(asset.getStringArray("string_array_property"));
             Assertions.assertNotNull(stringArray);
 
             for (int i = 0; i < 3; i++) {
@@ -133,7 +136,7 @@ public class KJsonTransformerBasedAssetLoaderPositiveTests extends KStandardTest
     @Test
     public void testAddAssetTypeDefinitionThatIsNotRegistered() {
 
-        KAssetLoader assetLoader = new KJsonTransformerBasedAssetLoader(
+        KAssetLoader assetLoader = new KJsonAssetLoader(
             new KStandardResourceLoader(
                 List.of(new KClasspathProtocol(
                     ClassLoader.getSystemClassLoader()
@@ -141,26 +144,6 @@ public class KJsonTransformerBasedAssetLoaderPositiveTests extends KStandardTest
             ),
             this.jsonParser,
             new String[] { "classpath:new_assets/" }
-//            new KJsonTransformerBasedAssetLoader.AssetTypeData(
-//                "type_1",
-//                Map.of(
-//                    "alias_1",
-//                    (v) -> {
-//                        Map<String, Object> data = new HashMap<>();
-//                        data.put("int_property", v.getInt("int_property"));
-//                        data.put("float_property", v.getFloat("float_property"));
-//                        data.put("boolean_property", v.getBoolean("boolean_property"));
-//                        data.put("string_property", v.getString("string_property"));
-//                        data.put("subdef_property", v.getSubdefinition("subdef_property"));
-//                        data.put("int_array_property", v.getIntArray("int_array_property"));
-//                        data.put("float_array_property", v.getFloatArray("float_array_property"));
-//                        data.put("boolean_array_property", v.getBooleanArray("boolean_array_property"));
-//                        data.put("string_array_property", v.getStringArray("string_array_property"));
-//                        data.put("subdef_array_property", v.getSubdefinitionArray("subdef_array_property"));
-//                        return new KMapAssetDefinition(data);
-//                    }
-//                )
-//            )
         );
 
         try {
@@ -186,7 +169,7 @@ public class KJsonTransformerBasedAssetLoaderPositiveTests extends KStandardTest
     @Test
     public void testAddNewAsset() {
 
-        KAssetLoader assetLoader = new KJsonTransformerBasedAssetLoader(
+        KAssetLoader assetLoader = new KJsonAssetLoader(
             new KStandardResourceLoader(
                 List.of(new KClasspathProtocol(
                     ClassLoader.getSystemClassLoader()
@@ -217,7 +200,7 @@ public class KJsonTransformerBasedAssetLoaderPositiveTests extends KStandardTest
         );
 
         try {
-            assetLoader.addAssetTypedef(new KJsonTransformerBasedAssetLoaderPositiveTests.Alias1Typedef());
+            assetLoader.addAssetTypedef(new KJsonAssetLoaderPositiveTests.Alias1Typedef());
 
             KJsonValue addedDef = this.jsonParser.parse("""
                 {
