+++ implementations/implementations-std-core/src/test/resources/new_assets_with_unknown_type/asset_1.json
@@ -1,39 +1,41 @@
 {
-    "type": "typpi_1",
+    "type": "type_1",
     "asset_id": "type_1.asset_1",
-    "int_property": 1,
-    "float_property": 1.0,
-    "boolean_property": false,
-    "string_property": "1234",
-    "subdef_property": {
-        "prop_1": "string"
-    },
-    "int_array_property": [
-        1,
-        2,
-        3
-    ],
-    "float_array_property": [
-        1.0,
-        2.0,
-        3.0
-    ],
-    "boolean_array_property": [
-        true,
-        true,
-        true
-    ],
-    "string_array_property": [
-        "1234",
-        "1235",
-        "9999"
-    ],
-    "subdef_array_property": [
-        {
+    "schema_version": 1,
+    "data": {
+        "float_property": 1.0,
+        "boolean_property": false,
+        "string_property": "1234",
+        "subdef_property": {
             "prop_1": "string"
         },
-        {
-            "prop_1": "string"
-        }
-    ]
+        "int_array_property": [
+            1,
+            2,
+            3
+        ],
+        "float_array_property": [
+            1.0,
+            2.0,
+            3.0
+        ],
+        "boolean_array_property": [
+            true,
+            true,
+            true
+        ],
+        "string_array_property": [
+            "1234",
+            "1235",
+            "9999"
+        ],
+        "subdef_array_property": [
+            {
+                "prop_1": "string"
+            },
+            {
+                "prop_1": "string"
+            }
+        ]
+    }
 }
