+++ implementations/implementations-std-entity/src/test/java/io/github/darthakiranihil/konna/entity/KStandardEntityFactoryPositiveTests.java
@@ -24,7 +24,7 @@ import io.github.darthakiranihil.konna.core.data.json.KStandardJsonParser;
 import io.github.darthakiranihil.konna.core.data.json.KStandardJsonTokenizer;
 import io.github.darthakiranihil.konna.core.di.KAppContainer;
 import io.github.darthakiranihil.konna.core.di.KEngineModule;
-import io.github.darthakiranihil.konna.core.io.KJsonTransformerBasedAssetLoader;
+import io.github.darthakiranihil.konna.core.io.KJsonAssetLoader;
 import io.github.darthakiranihil.konna.core.util.KReflectionUtils;
 import io.github.darthakiranihil.konna.entity.asset.KEntityMetadataCollection;
 import io.github.darthakiranihil.konna.entity.impl.TestBehaviour;
@@ -58,7 +58,7 @@ public class KStandardEntityFactoryPositiveTests extends KStandardTestClass {
                 new KSystemFeatures()
             )
         );
-        var assetLoader = new KJsonTransformerBasedAssetLoader(
+        var assetLoader = new KJsonAssetLoader(
             engineModule.resourceLoader(),
             new KStandardJsonParser(new KStandardJsonTokenizer()),
             new String[] { "classpath:assets/" }
